+++ NEWS.md
@@ -1,3 +1,16 @@
+# Version 0.68.0 [2025-04-18]
+
+## Documentation
+
+ * Fix minor help-page issues.
+
+## Bug Fixes
+
+ * `segmentByCBS()` had an internal bug which triggered an error on
+  `could not find function "nbrOfSegments"`, which began with
+  **future** 1.40.0.
+ 
+
 # Version 0.67.0 [2024-02-16]
 
 ## Documentation